# WIP: Cythonize away some perf hot spots