--- a/cuda_core/setup.py
+++ b/cuda_core/setup.py
@@ -2,33 +2,28 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+import glob
 import os
 
 from Cython.Build import cythonize
 from setuptools import Extension, setup
 from setuptools.command.build_ext import build_ext as _build_ext
 
-ext_modules = (
+
+# It seems setuptools' wildcard support has problems for namespace packages,
+# so we explicitly spell out all Extension instances.
+root_module = "cuda.core.experimental"
+root_path = f"{os.path.sep}".join(root_module.split(".")) + os.path.sep
+ext_files = glob.glob(f"{root_path}/**/*.pyx", recursive=True)
+def strip_prefix_suffix(filename):
+    return filename[len(root_path):-4]
+module_names = (strip_prefix_suffix(f) for f in ext_files)
+ext_modules = tuple(
     Extension(
-        "cuda.core.experimental._event",
-        sources=["cuda/core/experimental/_event.pyx"],
+        f"cuda.core.experimental.{mod.replace(os.path.sep, '.')}",
+        sources=[f"cuda/core/experimental/{mod}.pyx"],
         language="c++",
-    ),
-    Extension(
-        "cuda.core.experimental._dlpack",
-        sources=["cuda/core/experimental/_dlpack.pyx"],
-        language="c++",
-    ),
-    Extension(
-        "cuda.core.experimental._memoryview",
-        sources=["cuda/core/experimental/_memoryview.pyx"],
-        language="c++",
-    ),
-    Extension(
-        "cuda.core.experimental._kernel_arg_handler",
-        sources=["cuda/core/experimental/_kernel_arg_handler.pyx"],
-        language="c++",
-    ),
+    ) for mod in module_names
 )